% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAToNumeric.R
\name{convert_all_numericNA_to_median}
\alias{convert_all_numericNA_to_median}
\title{Convert NAs in a tribble to median of that column}
\usage{
convert_all_numericNA_to_median(data)
}
\arguments{
\item{data}{A tribble}
}
\value{
A tribble without NAs for numeric columns
}
\description{
Convert NAs in all numeric columns in a tribble to median of that column
}
\examples{
library(tidyverse)
library(datateachr)
data <- tribble(
~name,            ~birth_weight,  ~birth_year, 
"Luke Skywalker",  1.35      ,   1998        , 
"C-3PO"         ,  NA        ,   2000        ,  
"R2-D2"         ,  2.25      ,   NA        ,  
"Darth Vader"   ,  2.70      ,   NA        ,  
)
convert_all_numericNA_to_median(data)
}
